// src/app/shared/models/TimeClockPlus/employee-leave-report-request.ts
export class EmployeeLeaveReport {
    leaveYearId?: number | null;
    departmentId?: number | null;
    designationId?: number | null;
    employeeTypeId?: number | null;
    employeeCode?: string | null;
    leaveTypeId?: number | null;
    leaveStatus?: string | null;
    approvalLevelId?: number | null;
    fromDate?: string | Date | null;
    toDate?: string | Date | null;
}